% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autochart.R
\name{autochart}
\alias{autochart}
\title{Create an auto-generated chart from a data file.}
\usage{
autochart(file_path, x_var, y_var)
}
\arguments{
\item{file_path}{A character string specifying the path to the data file (.csv or .xlsx).}

\item{x_var}{A character string specifying the name of the variable for the x-axis.}

\item{y_var}{A character string specifying the name of the variable for the y-axis.}
}
\value{
A ggplot2 plot object.
}
\description{
This function reads a .csv or .xlsx file,
and automatically generates a chart (scatter plot, bar plot, or stacked bar plot)
based on the data types of the specified variables.

Author: Arman Mahmud

Email: armanmahmud.du18@gmail.com

Website: https://sites.google.com/view/armanmahmud1
}
\examples{
# Assuming you have a file named "my_data.csv" with columns 'A' and 'B'
library(autochart)
autochart("my_data.csv", "A", "B")
}
